Removed old network_http and related data structure. Its function is now either in the addons/news_manager, or in Online::RequestManager. Also added leack check and copy-preventing to online/Request.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@15034 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/graphics/gpuparticles.cpp
@@ -517,7 +517,7 @@ void ParticleSystemProxy::simulateHeightmap()
 
 void ParticleSystemProxy::simulateNoHeightmap()
 {
-	int timediff = GUIEngine::getLatestDt() * 1000.;
+	int timediff = int(GUIEngine::getLatestDt() * 1000.f);
 	int active_count = getEmitter()->getMaxLifeTime() * getEmitter()->getMaxParticlesPerSecond() / 1000;
 	core::matrix4 matrix = getAbsoluteTransformation();
 	glUseProgram(SimpleSimulationShader::Program);
